requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dictdiffer==0.8.1
Flask==1.1.2
Flask-BasicAuth==0.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
pyaml==20.4.0
PyYAML==5.4.1
requests==2.25.1
urllib3==1.26.3
viptela==0.3.2
Werkzeug==1.0.1